+++ inc/object.h
@@ -6,7 +6,7 @@
 /*   By: jyim <jyim@student.42kl.edu.my>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/05 09:37:39 by jyim              #+#    #+#             */
-/*   Updated: 2023/07/05 09:39:05 by jyim             ###   ########.fr       */
+/*   Updated: 2023/07/22 16:39:25 by jyim             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,10 +15,10 @@
 
 typedef struct s_sphere
 {
-	float	a;
-	float	b;
-	float	c;
-	float	t1;
-	float	t2;
+	double	a;
+	double	b;
+	double	c;
+	double	t1;
+	double	t2;
 }				t_sphere;
 #endif
