+++ inc/vector3.h
@@ -6,7 +6,7 @@
 /*   By: jyim <jyim@student.42kl.edu.my>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/05 10:28:55 by jyim              #+#    #+#             */
-/*   Updated: 2023/07/05 10:36:19 by jyim             ###   ########.fr       */
+/*   Updated: 2023/07/22 16:39:25 by jyim             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,12 +18,12 @@
 
 typedef struct s_vec3
 {
-	float	x;
-	float	y;
-	float	z;
+	double	x;
+	double	y;
+	double	z;
 }				t_vec3;
 
-t_vec3	*create_vec3(float _x, float _y, float _z);
+t_vec3	*create_vec3(double _x, double _y, double _z);
 t_vec3	add_vec3(t_vec3 vector1, t_vec3 vector2);
