+++ src/vector3.c
@@ -6,13 +6,13 @@
 /*   By: jyim <jyim@student.42kl.edu.my>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/05 10:17:04 by jyim              #+#    #+#             */
-/*   Updated: 2023/07/05 10:37:51 by jyim             ###   ########.fr       */
+/*   Updated: 2023/07/22 16:39:25 by jyim             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "vector3.h"
 
-t_vec3	*create_vec3(float _x, float _y, float _z)
+t_vec3	*create_vec3(double _x, double _y, double _z)
 {
 	t_vec3	*new_vec3;
 
@@ -43,9 +43,9 @@ t_vec3	sub_vec3(t_vec3 vec1, t_vec3 vec2)
 	return (result);
 }
 
-float	dot_vec3(t_vec3 vec1, t_vec3 vec2)
+double	dot_vec3(t_vec3 vec1, t_vec3 vec2)
 {
-	float	result;
+	double	result;
 
 	result = vec1.x * vec2.x + vec1.y * vec2.y + vec1.z * vec2.z;
 	return (result);
@@ -64,7 +64,7 @@ t_vec3	cross_vec3(t_vec3 vec1, t_vec3 vec2)
 t_vec3	*pnormalize(t_vec3 vec)
 {
 	t_vec3	*ret;
-	float	w;
+	double	w;
 
 	ret = malloc (sizeof(t_vec3));
 	if (!ret)
